Refactor: ts마이그레이션(auth레포지토리)
